requirements.txt
alabaster==0.7.8
alembic==1.0.10
anaconda-client==1.4.0
anaconda-navigator==1.8.4
anaconda-project==0.8.2
apptools==4.4.0
argcomplete==1.0.0
astroid==2.0.4
astropy==1.2.1
Babel==2.3.3
backports.csv==1.0.7
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.4.1
bitarray==0.8.1
blaze==0.10.1
blinker==1.4
bokeh==0.12.0
boto==2.40.0
Bottleneck==1.1.0
certifi==2018.8.24
cffi==1.6.0
chardet==3.0.4
chest==0.2.3
click==6.6
cloudpickle==0.2.1
clyent==1.2.2
colorama==0.3.7
comtypes==1.1.2
cond-calculation==1.0
conda==4.5.11
conda-build==1.21.3
configobj==5.0.6
contextlib2==0.5.3
cryptography==1.4
cx-Freeze==5.1
cycler==0.10.0
Cython==0.24
cytoolz==0.8.0
dask==0.10.0
datashape==0.5.2
decorator==4.0.10
defusedxml==0.6.0
descartes==1.1.0
dill==0.2.5
docutils==0.12
dominate==2.3.5
drawnow==0.71.2
dynd===c328ab7
elasticsearch==7.0.2
envisage==4.5.1
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==2.0.7
Flask==0.11.1
Flask-Admin==1.5.3
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==2.1.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.8.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gevent==1.1.1
greenlet==0.4.10
guess-language-spirit==0.5.3
h5py==2.6.0
HeapDict==1.0.0
idna==2.7
imagesize==0.7.1
ipykernel==4.3.1
ipython==4.2.0
ipython-genutils==0.1.0
ipywidgets==4.1.1
isort==4.3.4
itsdangerous==0.24
jdcal==1.2
jedi==0.9.0
Jinja2==2.10
jsonify==0.5
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.3.0
jupyter-console==4.1.1
jupyter-core==4.1.0
keyring==13.2.1
lazy-object-proxy==1.3.1
llvmlite==0.11.0
locket==0.2.0
lxml==3.6.0
Mako==1.0.12
MarkupSafe==0.23
matplotlib==1.5.1
mayavi==4.5.0
mccabe==0.6.1
menuinst==1.4.1
mistune==0.7.2
mpmath==0.19
multipledispatch==0.4.8
nb-anacondacloud==1.1.0
nb-conda==1.1.0
nb-conda-kernels==1.0.3
nbconvert==4.2.0
nbformat==4.0.1
nbpresent==3.0.2
networkx==1.11
nltk==3.2.1
nose==1.3.7
notebook==4.2.1
numba==0.26.0
numexpr==2.6.0
numpy==1.13.1
numpydoc==0.9.1
odfpy==1.4.0
odo==0.5.0
olefile==0.46
openpyxl==2.6.2
pandas==0.18.1
partd==0.3.4
path.py==0.0.0
pathlib2==2.1.0
patsy==0.4.1
pep8==1.7.0
pickleshare==0.7.2
Pillow==5.2.0
ply==3.8
proppant-allocation==1.0
psutil==4.3.0
py==1.4.31
pyasn1==0.1.9
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.14
pycrypto==2.6.1
pycurl==7.43.0.2
pyface==5.1.0
pyflakes==1.2.3
Pygments==2.1.3
PyGreSQL==5.0.2
PyJWT==1.7.1
pylint==2.1.1
PyOpenGL==3.1.1a1
pyOpenSSL==16.2.0
pyparsing==2.1.4
pyqtgraph==0.9.10
pyreadline==2.1
PySocks==1.6.8
pytest==2.9.2
python-dateutil==2.5.3
python-editor==1.0.4
pytz==2016.4
pywin32==220
PyYAML==3.11
pyzmq==15.2.0
QtAwesome==0.5.7
qtconsole==4.5.1
QtPy==1.7.1
requests==2.19.1
rope-py3k==0.9.4.post1
ruamel-yaml===-VERSION
scikit-image==0.12.3
scikit-learn==0.17.1
scipy==0.17.1
Shapely==1.5.17
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
sockjs-tornado==1.0.3
Sphinx==1.3.1
sphinx-rtd-theme==0.1.9
spyder==3.3.1
spyder-kernels==0.2.4
SQLAlchemy==1.0.13
statsmodels==0.6.1
sympy==1.1.1
tables==3.2.2
tablib==0.13.0
toolz==0.8.0
tornado==4.3
traitlets==4.2.1
traits==4.5.0
traitsui==5.1.0
triangle==20170429
typed-ast==1.1.0
unicodecsv==0.14.1
urllib3==1.23
visitor==0.1.3
vispy==0.4.0
Werkzeug==0.11.10
win-inet-pton==1.0.1
wrapt==1.10.11
WTForms==2.2.1
xlrd==1.0.0
XlsxWriter==0.9.2
xlwings==0.7.2
xlwt==1.1.2
